Irene validations

+++ src/backend/django_src/calendars/views/calendars.py
@@ -1,4 +1,5 @@
 from rest_framework.permissions import IsAuthenticated
+from ..permissions import IsCalendarOwner
 from ..models.Calendar import Calendar
 from ..models.Member import Member
 from ..serializers import CalendarListSerializer, CalendarPUTSerializer
@@ -17,15 +18,16 @@ from rest_framework.views import APIView
 
 # EndPoint: /calendars/list/
 class CalendarList(APIView):
-    # permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated]
 
     def get(self, request):
-        # Get all calendars that the user created(owner=user)
+        """View all calendars of the user"""
         calendars = Calendar.objects.filter(owner=request.user)
         serializer = CalendarListSerializer(calendars, many=True)
         return Response(serializer.data)
     
-    def post(self, request):    
+    def post(self, request):
+        """Create a new calendar"""
         serializer = CalendarListSerializer(data=request.data)
         
         if serializer.is_valid():
@@ -41,15 +43,21 @@ class CalendarList(APIView):
     
 # EndPoint: /calendars/<int:calendar_id>/
 class CalendarDetail(APIView):
-    # permission_classes = [IsAuthenticated] # Only the owner can view, edit, and delete the calendar
+    permission_classes = [IsAuthenticated, IsCalendarOwner]
 
     def get(self, request, calendar_id):
+        """View a specific calendar's details"""
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         serializer = CalendarListSerializer(calendar)
         return Response(serializer.data)
 
     def put(self, request, calendar_id):
+        """Edit a specific calendar's details"""
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         serializer = CalendarPUTSerializer(calendar, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -57,16 +65,21 @@ class CalendarDetail(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, calendar_id):
+        """Delete a specific calendar"""
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         calendar.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
 # EndPoint: /calendars/<int:calendar_id>/remindAll
 class CalendarRemind(APIView):
-    # permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsCalendarOwner]
     
     def post(self, request, calendar_id):
+        """Remind all members of the calendar to submit their availability"""
         calendar = get_object_or_404(Calendar, pk=calendar_id)
+        self.check_object_permissions(request, calendar)
         
         members = Member.objects.filter(calendar=calendar)
         owner_name = request.user.first_name
